Deduplicate call kent artwork avatar selection
