+++ app/utils/call-kent-artwork.ts
@@ -1,6 +1,7 @@
-import { toBase64 } from './misc.tsx'
+import { images } from '#app/images.tsx'
+import { getOptionalTeam, toBase64 } from './misc.tsx'
 
-type CallKentEpisodeArtworkAvatar =
+export type CallKentEpisodeArtworkAvatar =
 	| { kind: 'fetch'; url: string }
 	| { kind: 'public'; publicId: string }
 
@@ -38,10 +39,40 @@ type CallKentEpisodeArtworkOptions = {
 }
 
 const DESIGN_SIZE = 3000
+const KODY_PROFILE_BY_TEAM = {
+	RED: images.kodyProfileRed,
+	BLUE: images.kodyProfileBlue,
+	YELLOW: images.kodyProfileYellow,
+	UNKNOWN: images.kodyProfileGray,
+} as const
 
 // Cloudinary needs double-encoding for `l_text:` payloads.
 function doubleEncode(s: string) {
 	return encodeURIComponent(encodeURIComponent(s))
+}
+
+export function getCallKentEpisodeArtworkAvatar({
+	isAnonymous,
+	team,
+	gravatarUrl,
+}: {
+	isAnonymous: boolean
+	team: string
+	gravatarUrl?: string | null
+}): CallKentEpisodeArtworkAvatar {
+	if (isAnonymous) {
+		return { kind: 'public', publicId: images.kodyProfileGray.id }
+	}
+
+	if (gravatarUrl) {
+		return { kind: 'fetch', url: gravatarUrl }
+	}
+
+	const teamKey = getOptionalTeam(team)
+	return {
+		kind: 'public',
+		publicId: KODY_PROFILE_BY_TEAM[teamKey].id,
+	}
 }
 
 /**
